+++ exposed-core/src/main/kotlin/org/jetbrains/exposed/sql/statements/StatementInterceptor.kt
@@ -8,11 +8,21 @@ interface StatementInterceptor {
     fun beforeExecution(transaction: Transaction, context: StatementContext) {}
     fun afterExecution(transaction: Transaction, contexts: List<StatementContext>, executedStatement: PreparedStatementApi) {}
 
-    fun beforeCommit(transaction: Transaction) {}
+    fun beforeCommit(transaction: Transaction) { beforeCommit() }
+
+    @Deprecated("using beforeCommit with transaction")
+    //@Deprecated("using beforeCommit with transaction", level = DeprecationLevel.ERROR) \\ next version
+    //@Deprecated("using beforeCommit with transaction", level = DeprecationLevel.HIDDEN) \\ next version, backward compatibility
+    fun beforeCommit() {}
+    fun afterCommit(transaction: Transaction) { afterCommit() }
+
+    @Deprecated("using afterCommit with transaction")
+    //@Deprecated("using afterRollback with transaction", level = DeprecationLevel.ERROR) \\ next version
+    //@Deprecated("using afterRollback with transaction", level = DeprecationLevel.HIDDEN) \\ next version, backward compatibility
     fun afterCommit() {}
 
     fun beforeRollback(transaction: Transaction) {}
-    fun afterRollback() {}
+    fun afterRollback(transaction: Transaction) {}
 
     fun keepUserDataInTransactionStoreOnCommit(userData: Map<Key<*>, Any?>): Map<Key<*>, Any?> = emptyMap()
 }