+++ exposed-dao/src/main/kotlin/org/jetbrains/exposed/dao/References.kt
@@ -167,6 +167,7 @@ private fun <ID : Comparable<ID>> List<Entity<ID>>.preloadRelations(
                 (refObject as Referrers<ID, Entity<ID>, *, Entity<*>, Any>).reference.let { refColumn ->
                     val refIds = this.map { it.run { refColumn.referee<Any>()!!.lookup() } }
                     refObject.factory.warmUpReferences(refIds, refColumn)
+                    storeReferenceCache(refColumn, prop)
                 }
             }
             is OptionalReferrers<*, *, *, *, *> -> {
@@ -223,6 +224,9 @@ private fun <ID : Comparable<ID>> List<Entity<ID>>.preloadRelations(
 
 fun <SRCID : Comparable<SRCID>, SRC : Entity<SRCID>, REF : Entity<*>> Iterable<SRC>.with(vararg relations: KProperty1<out REF, Any?>): Iterable<SRC> =
     toList().apply {
+        if (any { it.isNewEntity() }) {
+            TransactionManager.current().flushCache()
+        }
         preloadRelations(*relations)
     }