+++ exposed-jdbc/src/main/kotlin/org/jetbrains/exposed/sql/statements/jdbc/JdbcConnectionImpl.kt
@@ -40,7 +40,15 @@ class JdbcConnectionImpl(override val connection: Connection) : ExposedConnectio
         get() = connection.autoCommit
         set(value) { connection.autoCommit = value }
 
-    override var transactionIsolation: Int = connection.transactionIsolation
+    override var transactionIsolation: Int = -1
+        get() {
+            if (field == -1) {
+                synchronized(this) {
+                    field = connection.transactionIsolation
+                }
+            }
+            return field
+        }
         set(value) {
             if (field != value) {
                 connection.transactionIsolation = value